requirements.txt
# Professional AI Stability Architecture
fastapi
uvicorn
python-multipart
python-dotenv
httpx
jinja2
# Industry Standard Embedding Engine (No Rust errors)
sentence-transformers
# Data & Vectors
faiss-cpu
sqlalchemy
psycopg2-binary
pgvector
anthropic
pypdf
tabulate
pandas
numpy
